Moved Repository Abstraction Files into their own namespace.
