+++ MadBand.WebApp/Models/Data/Context/Repositories/Abstraction/IIdentifiable.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MadBand.WebApp.Models.Data.Context.Repositories
+namespace MadBand.WebApp.Models.Data.Context.Repositories.Abstraction
 {
 	public interface IIdentifiable
 	{
