+++ MadBand.WebApp/Models/Data/Context/Repositories/Abstraction/ILookupByString.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MadBand.WebApp.Models.Data.Context.Repositories
+namespace MadBand.WebApp.Models.Data.Context.Repositories.Abstraction
 {
 public	interface ILookupByString<TEntity>:IRepository<TEntity>
 	{
